--- backstop.json
@@ -1,33 +1,0 @@
-{
-  "id": "backstop_default",
-  "viewports": [
-    {
-      "label": "MacBook Air",
-      "width": 1440,
-      "height": 900
-    }
-  ],
-  "onBeforeScript": "casper/onBefore.js",
-  "onReadyScript": "casper/onReady.js",
-  "scenarios": [
-    {
-      "label": "Index",
-      "url": "./dist/index.html",
-      "readySelector": ".smile",
-      "delay": 2000
-    }
-  ],
-  "paths": {
-    "bitmaps_reference": "backstop_data/bitmaps_reference",
-    "bitmaps_test": "backstop_data/bitmaps_test",
-    "engine_scripts": "backstop_data/engine_scripts",
-    "html_report": "backstop_data/html_report",
-    "ci_report": "backstop_data/ci_report"
-  },
-  "report": ["browser", "CI"],
-  "engine": "phantomjs",
-  "asyncCaptureLimit": 1,
-  "asyncCompareLimit": 50,
-  "debug": true,
-  "debugWindow": true
-}
