--- results.xml
@@ -1,1 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><testsuites time="0"><testsuite name="Index shows the placeholder by default" tests="0" failures="0" errors="0" time="0" timestamp="2017-12-09T19:27:11.209Z" package="./screenshot-test/index.screenshot.test"><system-out><![CDATA[]]></system-out></testsuite><testsuite name="Index shows the placeholder if there are no smiles" tests="0" failures="0" errors="0" time="0" timestamp="2017-12-09T19:27:11.209Z" package="./screenshot-test/index.screenshot.test"><system-out><![CDATA[]]></system-out></testsuite><testsuite name="Index shows the placeholder if the service is down" tests="0" failures="0" errors="0" time="0" timestamp="2017-12-09T19:27:11.209Z" package="./screenshot-test/index.screenshot.test"><system-out><![CDATA[]]></system-out></testsuite><testsuite name="Index shows the placeholder if the service fails" tests="0" failures="0" errors="0" time="0" timestamp="2017-12-09T19:27:11.210Z" package="./screenshot-test/index.screenshot.test"><system-out><![CDATA[]]></system-out></testsuite><testsuite name="Index shows the smile if there is just one" tests="0" failures="0" errors="0" time="0" timestamp="2017-12-09T19:27:11.210Z" package="./screenshot-test/index.screenshot.test"><system-out><![CDATA[]]></system-out></testsuite><testsuite name="Index shows the first smile with picture if there is more than one" tests="0" failures="0" errors="0" time="0" timestamp="2017-12-09T19:27:11.210Z" package="./screenshot-test/index.screenshot.test"><system-out><![CDATA[]]></system-out></testsuite></testsuites>
